Addressed homepage comments
