+++ components/EditScreenInfo/index.tsx
@@ -1,9 +1,10 @@
 import React from "react";
-import { StyleSheet } from "react-native";
 
-import { ExternalLink } from "./ExternalLink";
-import { MonoText } from "./StyledText";
-import { Text, View } from "./Themed";
+import { ExternalLink } from "../ExternalLink";
+import { MonoText } from "../StyledText";
+import { Text, View } from "../Themed";
+
+import { styles } from "./styles";
 
 import Colors from "@constants/Colors";
 
@@ -51,33 +52,3 @@ export default function EditScreenInfo({ path }: { path: string }) {
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  getStartedContainer: {
-    alignItems: "center",
-    marginHorizontal: 50,
-  },
-  homeScreenFilename: {
-    marginVertical: 7,
-  },
-  codeHighlightContainer: {
-    borderRadius: 3,
-    paddingHorizontal: 4,
-  },
-  getStartedText: {
-    fontSize: 17,
-    lineHeight: 24,
-    textAlign: "center",
-  },
-  helpContainer: {
-    marginTop: 15,
-    marginHorizontal: 20,
-    alignItems: "center",
-  },
-  helpLink: {
-    paddingVertical: 15,
-  },
-  helpLinkText: {
-    textAlign: "center",
-  },
-});
