+++ components/EditScreenInfo/styles.ts
@@ -1,0 +1,31 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  getStartedContainer: {
+    alignItems: "center",
+    marginHorizontal: 50,
+  },
+  homeScreenFilename: {
+    marginVertical: 7,
+  },
+  codeHighlightContainer: {
+    borderRadius: 3,
+    paddingHorizontal: 4,
+  },
+  getStartedText: {
+    fontSize: 17,
+    lineHeight: 24,
+    textAlign: "center",
+  },
+  helpContainer: {
+    marginTop: 15,
+    marginHorizontal: 20,
+    alignItems: "center",
+  },
+  helpLink: {
+    paddingVertical: 15,
+  },
+  helpLinkText: {
+    textAlign: "center",
+  },
+});
