+++ components/FunctionButton/styles.ts
@@ -1,0 +1,29 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+    blobs: {
+        width: 187.36,
+        height: 249.13,
+        backgroundColor: "white",
+        borderWidth: 1.5,
+        borderColor: "#ccc",
+        borderRadius: 30,
+    },
+    icons: {
+        marginLeft: 20,
+        marginTop: 20,
+    },
+    title: {
+        color: "black",
+        fontSize: 18,
+        fontWeight: "bold",
+        marginTop: 20,
+        marginLeft: 20,
+    },
+    stats: {
+        color: "gray",
+        fontSize: 16,
+        marginTop: 20,
+        marginLeft: 20, 
+    },
+});
