+++ components/GoalsProgress/index.tsx
@@ -1,6 +1,8 @@
 import { useEffect, useRef } from "react";
-import { View, Text } from "./Themed";
-import { Animated, StyleSheet } from "react-native";
+import { View, Text } from "../Themed";
+import { Animated } from "react-native";
+
+import { styles } from "./styles";
 
 import Target from "@assets/images/target";
 
@@ -15,7 +17,18 @@ interface Props {
 }
 
 export function GoalsProgress(props: Props) {
-    const { progress } = props;
+    let { progress } = props;
+
+    // Sanitize input to ensure current and goal values are non-negative
+    progress = progress.map(goal => {
+        const validCurr = Math.max(goal.curr, 0);
+        const validGoal = Math.max(goal.goal, 0);
+        return {
+            ...goal,
+            curr: validCurr,
+            goal: Math.max(validGoal, validCurr),
+        }
+    });
 
     // Create an array of Animated.Values, one for each goal
     const animatedValues = useRef<Animated.Value[]>(
@@ -75,51 +88,3 @@ export function GoalsProgress(props: Props) {
         </View>
     );
 }
-
-const styles = StyleSheet.create({
-    blob: {
-        width: 390.7,
-        height: 255.1,
-        backgroundColor: "white",
-        borderWidth: 1.5,
-        borderColor: "#ccc",
-        borderRadius: 30,
-    },
-    target: {
-        width: 40,
-        height: 40,
-        marginTop: 24,
-        marginLeft: 20,
-        backgroundColor: "transparent",
-    },
-    title: {
-        color: "black",
-        fontWeight: "bold",
-        fontSize: 20,
-        marginTop: 20,
-    },
-    barCaption: {
-        flexDirection: "row", 
-        justifyContent: "space-between", 
-        backgroundColor: "transparent",
-        marginHorizontal: 20,
-    },
-    info: {
-        color: "gray",
-        fontSize: 16,
-        marginTop: 20,
-    },
-    bar: {
-        width: 350.7,
-        height: 10,
-        borderRadius: 20,
-        marginTop: 5,
-        marginLeft: 20,
-        backgroundColor: "#e2cacf",
-    },
-    progress: {
-        height: 10, 
-        borderRadius: 20, 
-        backgroundColor: "#7A0019",
-    },
-});
