+++ components/StatsBlob/styles.ts
@@ -1,0 +1,45 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+    blobs: {
+        width: 187.36,
+        height: 249.13,
+        backgroundColor: "white",
+        borderWidth: 1.5,
+        borderColor: "#ccc",
+        borderRadius: 30,
+    },
+    icons: {
+        marginLeft: 20,
+        marginTop: 20,
+    },
+    titles: {
+        color: "gray",
+        fontSize: 16,
+        marginTop: 20,
+        marginLeft: 20,
+    },
+    stats: {
+        color: "gray",
+        fontSize: 16,
+        marginLeft: 20, 
+    },
+    current: {
+        color: "black",
+        fontSize: 24,
+        fontWeight: "bold",
+    },
+    bar: {
+        width: 150,
+        height: 10,
+        borderRadius: 20,
+        marginLeft: 20, 
+        marginTop: 25,
+        backgroundColor: "#e2cacf",
+    },
+    progress: {
+        height: 10,
+        borderRadius: 20,
+        backgroundColor: "#7A0019"
+    }
+});
