+++ components/Themed/index.tsx
@@ -6,7 +6,7 @@
 import { Text as DefaultText, View as DefaultView } from "react-native";
 
 import Colors from "@constants/Colors";
-import { useColorScheme } from "./useColorScheme";
+import { useColorScheme } from "../useColorScheme";
 
 type ThemeProps = {
   lightColor?: string;
